test: controller 테스트 추가
- 기존의 MissionStepTest 의 API 테스트를 각각의 controller 테스트로 이동
