+++ src/test/java/roomescape/controller/reservation/ReservationControllerTest.java
@@ -1,0 +1,98 @@
+package roomescape.controller.reservation;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import io.restassured.RestAssured;
+import io.restassured.http.ContentType;
+import java.time.LocalTime;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.test.annotation.DirtiesContext;
+import roomescape.controller.reservationtime.ReservationTimeController;
+import roomescape.controller.reservationtime.request.ReservationTimeRequest;
+import roomescape.model.Reservation;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+class ReservationControllerTest {
+
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
+
+    @Autowired
+    private ReservationTimeController reservationTimeController;
+
+    @DisplayName("예약을 추가할 수 있다.")
+    @Test
+    void saveReservation() {
+        //given
+        Map<String, Object> reservation = creatReservationFixture();
+
+        //when
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(reservation)
+                .when().post("/reservations")
+                .then().log().all()
+                .statusCode(200);
+
+        //then
+        Integer count = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
+        assertThat(count).isEqualTo(1);
+    }
+
+    @DisplayName("예약을 조회 할 수 있다.")
+    @Test
+    void readAllReservation() {
+        //given
+        saveReservation();
+
+        //when
+        List<Reservation> reservations = RestAssured.given().log().all()
+                .when().get("/reservations")
+                .then().log().all()
+                .statusCode(200).extract()
+                .jsonPath().getList(".", Reservation.class);
+
+        //then
+        Integer count = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
+
+        assertThat(reservations.size()).isEqualTo(count);
+    }
+
+    @DisplayName("예약번호에 따른 예약 정보를 삭제할 수 있다.")
+    @Test
+    void deleteReservation() {
+        //given
+        saveReservation();
+
+        //when
+        RestAssured.given().log().all()
+                .when().delete("/reservations/1")
+                .then().log().all()
+                .statusCode(200);
+
+        //then
+        Integer countAfterDelete = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
+        assertThat(countAfterDelete).isEqualTo(0);
+    }
+
+    private Map<String, Object> creatReservationFixture() {
+        saveReservationTimeFixture();
+        Map<String, Object> reservation = new HashMap<>();
+        reservation.put("name", "브라운");
+        reservation.put("date", "2023-08-05");
+        reservation.put("timeId", 1);
+        return reservation;
+    }
+
+    private void saveReservationTimeFixture() {
+        reservationTimeController.save(new ReservationTimeRequest(LocalTime.of(10, 0)));
+    }
+}
